+++ src/torchmetrics/functional/multimodal/__init__.py
@@ -12,11 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.functional.multimodal.lve import lip_vertex_error
-from torchmetrics.utilities.imports import _TRANSFORMERS_GREATER_EQUAL_4_10
+from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 __all__ = ["lip_vertex_error"]
 
-if _TRANSFORMERS_GREATER_EQUAL_4_10:
+if _TRANSFORMERS_AVAILABLE:
     from torchmetrics.functional.multimodal.clip_iqa import clip_image_quality_assessment
     from torchmetrics.functional.multimodal.clip_score import clip_score