+++ src/torchmetrics/text/__init__.py
@@ -25,7 +25,7 @@ from torchmetrics.text.ter import TranslationEditRate
 from torchmetrics.text.wer import WordErrorRate
 from torchmetrics.text.wil import WordInfoLost
 from torchmetrics.text.wip import WordInfoPreserved
-from torchmetrics.utilities.imports import _TRANSFORMERS_GREATER_EQUAL_4_4
+from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 __all__ = [
     "BLEUScore",
@@ -44,7 +44,7 @@ __all__ = [
     "WordInfoPreserved",
 ]
 
-if _TRANSFORMERS_GREATER_EQUAL_4_4:
+if _TRANSFORMERS_AVAILABLE:
     from torchmetrics.text.bert import BERTScore
     from torchmetrics.text.infolm import InfoLM